Change type declaration for used and total in MemoryManager from uint32_t to uint64_t for improved memory handling

+++ Kernel/memoryManagement/dummyMemoryManager.c
@@ -6,8 +6,8 @@
 
 typedef struct MemoryManager {
 	char *nextAddress;
-	uint32_t used;
-	uint32_t total;
+	uint64_t used;
+	uint64_t total;
 	memStatus status;
 } MemoryManager;
